findbyemail service  adoption_user relation, fix serializer issue  (cyclic)

+++ src/Controller/AdoptionController.php
@@ -4,22 +4,32 @@ namespace App\Controller;
 
 use App\Entity\Adoption;
 use App\Repository\AdoptionRepository;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\SerializerInterface;
 
 class AdoptionController extends AbstractFOSRestController
 {
 
     private $adoptionRepository;
     private $entityManager;
+    private $serializer;
+    private $userRepo;
 
-    public function __construct(AdoptionRepository $repository, EntityManagerInterface $em)
+    public function __construct(AdoptionRepository $repository, EntityManagerInterface $em, SerializerInterface $serializer, UserRepository $userRepo)
     {
         $this->adoptionRepository = $repository;
         $this->entityManager = $em;
+        $this->serializer = $serializer;
+        $this->userRepo = $userRepo;
     }
 
     /**
@@ -34,26 +44,27 @@ class AdoptionController extends AbstractFOSRestController
         $animal = $requst->query->get('animal');
         $createdAt = $requst->query->get('createdAt');
         $description = $requst->query->get('description');
+        $user_id = $requst->query->get('user_id');
 
-        if (isset($title) || isset($animal) || isset($description) || isset($createdAt)) {
-            $criteria = $this->createCriteria($title, $description, $createdAt, $animal);
+        if (isset($title) || isset($animal) || isset($description) || isset($createdAt) || isset($user_id)) {
+            $criteria = $this->createCriteria($title, $description, $createdAt, $animal, $user_id);
             if (!isset($page) && !isset($size)) {
                 $adoptions =  $this->adoptionRepository->findBy($criteria);
-                return $this->json($adoptions, Response::HTTP_OK);
+                return new Response($this->handleCircularReference($adoptions), Response::HTTP_OK);
             }
             $page = isset($page) && $page > 0 ? $page : 1;
             $offset = isset($size) ? ($page - 1) * $size : 0;
             $adoptions = $this->adoptionRepository->findBy($criteria, null, isset($size) ? $size :  8,  $offset);
-            return $this->json($adoptions, Response::HTTP_OK);
+            return new Response($this->handleCircularReference($adoptions), Response::HTTP_OK);
         }
 
         // if not paginated
         if (!isset($page) && !isset($size)) {
             $adoptions = $this->adoptionRepository->findAll();
-            return $this->json($adoptions, Response::HTTP_OK);
+            return new Response($this->handleCircularReference($adoptions), Response::HTTP_OK);
         }
         $adoptions = $this->adoptionRepository->findPaged($page, $size);
-        return $this->json($adoptions, Response::HTTP_OK);
+        return new Response($this->handleCircularReference($adoptions), Response::HTTP_OK);
     }
 
 
@@ -72,7 +83,10 @@ class AdoptionController extends AbstractFOSRestController
     public function findOne($id): Response
     {
         $adoption = $this->adoptionRepository->find($id);
-        return $this->json($adoption, Response::HTTP_OK);
+        if ($adoption == null) {
+            return new Response('Adoption not found', Response::HTTP_NOT_FOUND);
+        }
+        return new Response($this->handleCircularReference($adoption), Response::HTTP_OK);
     }
 
     /**
@@ -81,9 +95,12 @@ class AdoptionController extends AbstractFOSRestController
     public function delete($id): Response
     {
         $adoption = $this->adoptionRepository->find($id);
+        if ($adoption == null) {
+            return new Response('Adoption not found', Response::HTTP_NOT_FOUND);
+        }
         $this->entityManager->remove($adoption);
         $this->entityManager->flush();
-        return $this->json($adoption, Response::HTTP_OK);
+        return new Response($this->handleCircularReference($adoption), Response::HTTP_OK);
     }
 
     /**
@@ -93,10 +110,13 @@ class AdoptionController extends AbstractFOSRestController
     {
         $data = json_decode($request->getContent(), true);
         $adoption = $this->adoptionRepository->find($id);
+        if ($adoption == null) {
+            return new Response('Adoption not found', Response::HTTP_NOT_FOUND);
+        }
         $adoption = $this->adoptionDto($adoption, $data);
         $this->entityManager->persist($adoption);
         $this->entityManager->flush();
-        return $this->json($adoption);
+        return new Response($this->handleCircularReference($adoption), Response::HTTP_OK);
     }
 
     /**
@@ -108,18 +128,24 @@ class AdoptionController extends AbstractFOSRestController
         $adoption = $this->adoptionDto(new Adoption(), $data);
         $this->entityManager->persist($adoption);
         $this->entityManager->flush();
-        return $this->json($adoption, Response::HTTP_CREATED);
+        return new Response($this->handleCircularReference($adoption), Response::HTTP_CREATED);
     }
 
     private function adoptionDto(Adoption $adoption, $data)
     {
+        $user_id = $data['user']['id'];
+        if (isset($user_id)) {
+            $user = $this->userRepo->find((int) $user_id);
+            $adoption->setUser($user);
+        }
         $adoption->setTitle($data['title']);
         $adoption->setDescription($data['description']);
+        $adoption->setAnimal($data['animal']);
         $adoption->setAnimal($data['animal']);
         return $adoption;
     }
 
-    private function createCriteria($title, $description, $creationAt, $animal): array
+    private function createCriteria($title, $description, $creationAt, $animal, $user_id): array
     {
         $criteria = [];
         if (isset($title)) {
@@ -134,6 +160,20 @@ class AdoptionController extends AbstractFOSRestController
         if (isset($animal)) {
             $criteria['animal'] = $animal;
         }
+        if (isset($user_id)) {
+            $criteria['user'] = $user_id;
+        }
         return $criteria;
+    }
+
+    function handleCircularReference($objectToSerialize)
+    {
+        // Serialize your object in Json
+        $jsonObject = $this->serializer->serialize($objectToSerialize, 'json', [
+            'circular_reference_handler' => function ($object) {
+                return $object->getId();
+            }
+        ]);
+        return $jsonObject;
     }
 }
